FlaskPython/requirements.txt
flask
pandas
sqlalchemy
pymysql
joblib
scikit-learn
python-dotenv